requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
django-jazzmin==3.0.2
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.5